render_requirements.txt
discord.py==2.3.2
flask==3.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.32.4
gunicorn==21.2.0
werkzeug==3.0.1